feat: extract Box internals as a hook
