+++ lib/components/Box/Box.tsx
@@ -1,31 +1,7 @@
-import type { AllHTMLAttributes, ElementType, PropsWithChildren } from 'react';
+import type { PropsWithChildren } from 'react';
 import React, { forwardRef } from 'react';
 
-import {
-	elementStyles,
-	type ElementStylesProps,
-} from '../../styles/elementStyles';
-import { OdComponent, TestId } from '../../types';
-import { dataAttrs } from '../../utils/dataAttrs';
-
-/**
- * Use CommonBoxProps to help consistently define additional utility props of a component
- */
-export interface CommonBoxProps
-	extends OdComponent,
-		PropsWithChildren,
-		TestId {}
-
-export interface BoxProps
-	extends Omit<
-			AllHTMLAttributes<HTMLElement>,
-			'as' | 'className' | 'color' | 'height' | 'is' | 'size' | 'width'
-		>,
-		CommonBoxProps,
-		ElementStylesProps {
-	/** Pass in the HTML element or JSX component that should be rendered for the box */
-	as?: ElementType;
-}
+import { useBox, type UseBoxProps as BoxProps } from './useBox/useBox';
 
 /**
  * The Box component provides powerful responsive style props to a given tag, defaulting to a `<div>` element.
@@ -45,255 +21,12 @@ export interface BoxProps
  * @example
  * <Box as={<MyCustomThing />} borderColor="info" borderWidth="1" />
  */
-export const Box = forwardRef<HTMLElement, BoxProps>(
-	(
-		{
-			as = 'div',
-			children,
-			className,
-			odComponent,
-			testId,
-
-			// style props
-			display,
-			height,
-			maxWidth,
-			minWidth,
-			overflow,
-			overflowX,
-			overflowY,
-			pointerEvents,
-			position,
-			size,
-			userSelect,
-			useVar,
-			width,
-
-			bg,
-			backgroundColor,
-			backgroundColour,
-			boxShadow,
-			color, // modern semantic colour tokens
-			colour, // legacy colours
-			fg,
-			opacity,
-
-			fontSize,
-			fontWeight,
-			lineHeight,
-			text,
-			textAlign,
-			textWrap,
-
-			borderRadius,
-			borderColor,
-			borderStyle,
-			borderWidth,
-			borderWidthX,
-			borderWidthY,
-			borderBottomColor,
-			borderBottomStyle,
-			borderBottomWidth,
-			borderLeftColor,
-			borderLeftStyle,
-			borderLeftWidth,
-			borderRightColor,
-			borderRightStyle,
-			borderRightWidth,
-			borderTopColor,
-			borderTopStyle,
-			borderTopWidth,
-			borderWidthTop,
-			borderWidthRight,
-			borderWidthBottom,
-			borderWidthLeft,
-
-			borderColour,
-			borderColourX,
-			borderColourY,
-			borderBottomColour,
-			borderLeftColour,
-			borderRightColour,
-			borderTopColour,
-
-			alignContent,
-			alignItems,
-			alignSelf,
-			columnGap,
-			flexDirection,
-			flexGrow,
-			flexShrink,
-			flexWrap,
-			gap,
-			gridAutoColumns,
-			gridAutoFlow,
-			gridAutoRows,
-			gridColumns,
-			justifyContent,
-			placeItems,
-			rowGap,
-			order,
-
-			m,
-			mb,
-			ml,
-			mr,
-			mt,
-			mx,
-			my,
-			margin,
-			marginX,
-			marginY,
-			marginBottom,
-			marginLeft,
-			marginRight,
-			marginTop,
-
-			p,
-			pb,
-			pl,
-			pr,
-			pt,
-			px,
-			py,
-			padding,
-			paddingX,
-			paddingY,
-			paddingBottom,
-			paddingLeft,
-			paddingRight,
-			paddingTop,
-
-			...allOtherProps
-		},
-		ref,
-	) => {
-		const Component = as;
-
-		const allClasses = elementStyles({
-			as,
-			className,
-
-			display,
-			height,
-			maxWidth,
-			minWidth,
-			overflow,
-			overflowX,
-			overflowY,
-			pointerEvents,
-			position,
-			size,
-			userSelect,
-			useVar,
-			width,
-
-			bg,
-			backgroundColor,
-			backgroundColour,
-			boxShadow,
-			color,
-			colour,
-			fg,
-			opacity,
-
-			fontSize,
-			fontWeight,
-			lineHeight,
-			text,
-			textAlign,
-			textWrap,
-
-			borderRadius,
-			borderColor,
-			borderStyle,
-			borderWidth,
-			borderWidthX,
-			borderWidthY,
-			borderBottomColor,
-			borderBottomStyle,
-			borderBottomWidth,
-			borderLeftColor,
-			borderLeftStyle,
-			borderLeftWidth,
-			borderRightColor,
-			borderRightStyle,
-			borderRightWidth,
-			borderTopColor,
-			borderTopStyle,
-			borderTopWidth,
-			borderWidthTop,
-			borderWidthRight,
-			borderWidthBottom,
-			borderWidthLeft,
-
-			borderColour,
-			borderColourX,
-			borderColourY,
-			borderBottomColour,
-			borderLeftColour,
-			borderRightColour,
-			borderTopColour,
-
-			alignContent,
-			alignItems,
-			alignSelf,
-			columnGap,
-			flexDirection,
-			flexGrow,
-			flexShrink,
-			flexWrap,
-			gap,
-			gridAutoColumns,
-			gridAutoFlow,
-			gridAutoRows,
-			gridColumns,
-			justifyContent,
-			placeItems,
-			rowGap,
-			order,
-
-			m,
-			mb,
-			ml,
-			mr,
-			mt,
-			mx,
-			my,
-			margin,
-			marginX,
-			marginY,
-			marginBottom,
-			marginLeft,
-			marginRight,
-			marginTop,
-
-			p,
-			pb,
-			pl,
-			pr,
-			pt,
-			px,
-			py,
-			padding,
-			paddingX,
-			paddingY,
-			paddingBottom,
-			paddingLeft,
-			paddingRight,
-			paddingTop,
-		});
+export const Box = forwardRef<HTMLElement, PropsWithChildren<BoxProps>>(
+	({ children, ...props }, ref) => {
+		const { Component, componentProps } = useBox(props);
 
 		return (
-			<Component
-				{...allOtherProps}
-				{...dataAttrs({
-					'od-component': odComponent?.toLocaleLowerCase(),
-					testid: testId,
-				})}
-				className={allClasses}
-				ref={ref}
-			>
+			<Component {...componentProps} ref={ref}>
 				{children}
 			</Component>
 		);
@@ -301,3 +34,5 @@ export const Box = forwardRef<HTMLElement, BoxProps>(
 );
 
 Box.displayName = 'Box';
+
+export { type UseBoxProps as BoxProps } from './useBox/useBox';
