+++ lib/components/Box/index.ts
@@ -1,3 +1,2 @@
-export { Box, type BoxProps } from './Box';
-export { elementStyles as useBoxStyles } from '../../styles/elementStyles';
-export type { ElementStylesProps as BoxStyleProps } from '../../styles/elementStyles';
+export * from './Box';
+export * from './useBox/useBox';
